pipeline-sobel-fpga/src/requirements.txt
pyserial
Pillow
tqdm